about页面

+++ res/Btrade-mobile/handler/main.py
@@ -156,18 +156,18 @@ class AboutHandler(BaseHandler):
                 quoteinfo = self.db.get(
                     "select price from quote where purchaseinfoid=%s and state=1 order by price desc limit 0,1", item)
                 if purchaseinfo["unit"] == u"公斤":
-                    sum_quantity += int(purchaseinfo["quantity"]) / 1000
-                    sum_price += int(purchaseinfo["quantity"]) * float(quoteinfo["price"])
+                    sum_quantity += float(purchaseinfo["quantity"]) / 1000
+                    sum_price += float(purchaseinfo["quantity"]) * float(quoteinfo["price"])
 
                 elif purchaseinfo["unit"] == u"吨":
-                    sum_quantity += int(purchaseinfo["quantity"])
-                    sum_price += int(purchaseinfo["quantity"]) * float(quoteinfo["price"]) * 1000
+                    sum_quantity += float(purchaseinfo["quantity"])
+                    sum_price += float(purchaseinfo["quantity"]) * float(quoteinfo["price"]) * 1000
 
         user_count = self.db.execute_rowcount("select id from users where type not in(1,2,9)")
         supplier_count = self.db.execute_rowcount("select id from supplier where pushstatus!=2")
         total = user_count + supplier_count
         show_data = {"accept_company_num": accept_company_num, "accept_quote_user_num": accept_quote_user_num,"total":total,
-                         "accept_num": accept_num, "sum_quantity": sum_quantity, "sum_price": int(sum_price / 10000)}
+                         "accept_num": accept_num, "sum_quantity": int(sum_quantity), "sum_price": int(sum_price / 10000)}
         self.render("about.html",data=show_data)
 
 class CenterHandler(BaseHandler):
